@@ -1,4 +1,4 @@
-name: flutterapp
+name: myasteer
 description: "A new Flutter project."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -29,32 +29,45 @@ environment:
 # versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
+  
     sdk: flutter
+    
+    
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.8
-    equatable: ^2.0.7
-  hive_flutter: ^1.1.0
-  get_it: ^8.0.3
-  uuid: ^4.5.1
-  shared_preferences: ^2.3.5
-  path_provider: ^2.1.5
   dio: ^5.7.0
+  hive_flutter: ^1.1.0
   pretty_dio_logger: ^1.4.0
-  flutter_bloc: ^9.0.0
-  bloc: ^9.0.0
-  hive_generator: ^2.0.1
-  build_runner: ^2.4.14
+  shared_preferences: ^2.3.4
+  multi_select_flutter: ^4.1.3
+  uuid: ^4.5.1
   dartz: ^0.10.1
-  connectivity_plus: ^6.1.1
+  path_provider: ^2.1.5
+  equatable: ^2.0.7
   image_picker: ^1.1.2
+  awesome_dio_interceptor: ^1.3.0
+  json_annotation: ^4.9.0
+  connectivity_plus: ^6.1.1
   permission_handler: ^11.3.1
+  json_serializable: ^6.9.0
+  sensors_plus: ^6.1.1
+
+  get_it: ^8.0.3
+  flutter_bloc: ^8.1.6
+  bloc: ^8.1.4
+  smooth_page_indicator: ^1.2.0+3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+ 
+
+  build_runner: ^2.4.14
+  hive_generator: ^2.0.1
+  test_cov_console: ^0.2.2
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
@@ -76,8 +89,25 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/icon/
-    - assets/images/
+    - assets/images/Logo.png
+    - assets/images/facebook.png
+    - assets/images/google.png
+    - assets/images/onboarding1.png
+    - assets/images/onboarding2.png
+    - assets/images/logoAster.png
+    - assets/images/doctor.png
+    - assets/images/dummy.jpg
+    - assets/images/login-i.png
+    - assets/images/doc.png
+    - assets/icons/pediatric.png
+    - assets/icons/neurology.jpg
+    - assets/icons/Tooth.png
+    - assets/icons/cardi.png
+    - assets/icons/derma.jpg
+
+    
+    
+  
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
@@ -91,24 +121,11 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
   fonts:
-    - family: OpenSans ExtraBold
+    - family: Rockwell
       fonts:
-        - asset: assets/fonts/OpenSans_Condensed-ExtraBold.ttf
-    - family: OpenSans SemiBoldItalic
-      fonts:
-        - asset: assets/fonts/OpenSans_Condensed-SemiBoldItalic.ttf
-    - family: OpenSans Regular
-      fonts:
-        - asset: assets/fonts/OpenSans_SemiCondensed-Regular.ttf
-    - family: OpenSans Bold
-      fonts:
-        - asset: assets/fonts/OpenSans_SemiCondensed-Bold.ttf
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
+        - asset: assets/fonts/ROCK.TTF
+        - asset: assets/fonts/Rockwell-Bold.TTF
+          style: italic
   #   - family: Trajan Pro
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
