@@ -1,31 +1,122 @@
 import 'package:flutter/material.dart';
+import 'package:flutterapp/app/constants/theme_constant.dart';
+// import 'package:myAster-Mobile-App-main/app/constants/theme_constant.dart';
 
-ThemeData getApplicationTheme() {
+ThemeData getApplicationTheme({required bool isDarkMode}) {
   return ThemeData(
-    appBarTheme: const AppBarTheme(
-        color: Colors.white,
-        elevation: 4,
-        shadowColor: Colors.black,
-        titleTextStyle: TextStyle(
-          fontSize: 16,
-          color: Colors.black,
-          fontWeight: FontWeight.bold,
-        )),
-    primarySwatch: Colors.orange,
-    scaffoldBackgroundColor: Colors.grey[200],
-    fontFamily: 'Montserrat Bold',
+    colorScheme: isDarkMode
+        ? const ColorScheme.dark(primary: ThemeConstant.darkPrimaryColor)
+        : const ColorScheme.light(primary: ThemeConstant.lightPrimaryColor),
+    scaffoldBackgroundColor: isDarkMode
+        ? ThemeConstant.darkPrimaryColor
+        : ThemeConstant.lightPrimaryColor,
+    fontFamily: "IM_Fell_English_SC",
+    textSelectionTheme: TextSelectionThemeData(
+      cursorColor: isDarkMode
+          ? ThemeConstant.lightPrimaryColor
+          : ThemeConstant.darkPrimaryColor,
+      selectionColor: isDarkMode
+          ? ThemeConstant.lightPrimaryColor.withOpacity(0.4)
+          : ThemeConstant.darkPrimaryColor.withOpacity(0.4),
+      selectionHandleColor: isDarkMode
+          ? ThemeConstant.lightPrimaryColor
+          : ThemeConstant.darkPrimaryColor,
+    ),
     elevatedButtonTheme: ElevatedButtonThemeData(
-      style: ElevatedButton.styleFrom(
-        textStyle: const TextStyle(
-          fontSize: 18,
+        style: ElevatedButton.styleFrom(
+      textStyle: const TextStyle(
+          fontSize: 20,
           fontWeight: FontWeight.bold,
-          color: Colors.white,
-        ),
-        backgroundColor: Colors.orange,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(5),
+          fontFamily: "IM_FELL_Great_Primer"),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(9)),
+      foregroundColor: isDarkMode
+          ? ThemeConstant.darkPrimaryColor
+          : ThemeConstant.lightPrimaryColor,
+      backgroundColor: isDarkMode
+          ? ThemeConstant.lightPrimaryColor
+          : ThemeConstant.darkPrimaryColor,
+    )),
+    bottomNavigationBarTheme: BottomNavigationBarThemeData(
+      backgroundColor: isDarkMode
+          ? ThemeConstant.darkPrimaryColor
+          : ThemeConstant.lightPrimaryColor,
+      selectedItemColor: isDarkMode
+          ? ThemeConstant.lightPrimaryColor
+          : ThemeConstant.darkPrimaryColor,
+      unselectedItemColor: isDarkMode
+          ? Colors.white.withOpacity(0.7)
+          : Colors.black.withOpacity(0.7),
+      selectedIconTheme: IconThemeData(
+        size: 28, // Custom size for selected icons
+        color: isDarkMode
+            ? ThemeConstant.lightPrimaryColor
+            : ThemeConstant.darkPrimaryColor,
+      ),
+      unselectedIconTheme: IconThemeData(
+        size: 24, // Custom size for unselected icons
+        color: isDarkMode
+            ? Colors.white.withOpacity(0.7)
+            : Colors.black.withOpacity(0.7),
+      ),
+      elevation: 8.0,
+      showUnselectedLabels: true, // Show unselected labels
+      selectedLabelStyle: const TextStyle(
+        fontWeight: FontWeight.bold,
+        fontSize: 14, // Customize label size
+      ),
+    ),
+    inputDecorationTheme: InputDecorationTheme(
+        hintStyle: const TextStyle(color: Colors.grey, fontSize: 13),
+        contentPadding: const EdgeInsets.symmetric(horizontal: 20),
+        enabledBorder: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(11),
+            borderSide: BorderSide(
+                color: isDarkMode
+                    ? ThemeConstant.lightPrimaryColor
+                    : ThemeConstant.darkPrimaryColor,
+                width: 1)),
+        focusedBorder: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(11),
+            borderSide: BorderSide(
+                color: isDarkMode
+                    ? ThemeConstant.lightPrimaryColor
+                    : ThemeConstant.darkPrimaryColor,
+                width: 2)),
+        errorBorder: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(11),
+            borderSide: const BorderSide(color: Colors.red, width: 1)),
+        focusedErrorBorder: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(11),
+            borderSide: const BorderSide(color: Colors.red, width: 2))),
+    appBarTheme: AppBarTheme(
+      backgroundColor: isDarkMode
+          ? ThemeConstant.darkPrimaryColor
+          : ThemeConstant.lightPrimaryColor,
+      elevation: 0,
+      centerTitle: true,
+      iconTheme: IconThemeData(
+        color: isDarkMode
+            ? ThemeConstant.lightPrimaryColor
+            : ThemeConstant.darkPrimaryColor,
+      ),
+      actionsIconTheme: IconThemeData(
+        color: isDarkMode
+            ? ThemeConstant.lightPrimaryColor
+            : ThemeConstant.darkPrimaryColor,
+      ),
+    ),
+    cardTheme: CardTheme(
+      color: isDarkMode ? const Color(0xFF171717) : Colors.white,
+      elevation: 4, // Controls the shadow depth
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(5), // Rounded corners
+        side: const BorderSide(
+          color: Color(0xFFE1E1D5),
+          width: 0.4, // Border width
         ),
       ),
+      // margin: const EdgeInsets.symmetric(
+      //     horizontal: 20, vertical: 10), // Default margins
     ),
   );
 }
